backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
requests
python-multipart
apscheduler
aiohttp
paramiko
cryptography